Extracting Money and Product classes
