+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Discount.java
@@ -1,0 +1,24 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.math.BigDecimal;
+
+public class Discount {
+    private String discountCause;
+    private BigDecimal discount;
+
+    public String getDiscountCause() {
+        return discountCause;
+    }
+
+    public void setDiscountCause(String discountCause) {
+        this.discountCause = discountCause;
+    }
+
+    public BigDecimal getDiscount() {
+        return discount;
+    }
+
+    public void setDiscount(BigDecimal discount) {
+        this.discount = discount;
+    }
+}
