+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Money.java
@@ -1,0 +1,6 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+public class Money {
+    private int value;
+    private String currency;
+}
