+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/OfferItem.java
@@ -18,87 +18,98 @@ import java.util.Date;
 public class OfferItem {
 
     // product
-    private String productId;
+//    private String productId;
 
-    private BigDecimal productPrice;
+//    private BigDecimal productPrice;
 
-    private String productName;
+//    private String productName;
 
-    private Date productSnapshotDate;
+//    private Date productSnapshotDate;
 
-    private String productType;
+//    private String productType;
 
     private int quantity;
 
     private BigDecimal totalCost;
 
-    private String currency;
+//    private String currency;
 
-    // discount
-    private String discountCause;
+    private Discount discount;
 
-    private BigDecimal discount;
+    private Money money;
+
+    public Discount getDiscount() {
+        return discount;
+    }
+
+    public void setDiscount(Discount discount) {
+        this.discount = discount;
+    }
+// discount
+//    private String discountCause;
+//
+//    private BigDecimal discount;
+
+//    public OfferItem(String productId, BigDecimal productPrice, String productName, Date productSnapshotDate,
+//            String productType, int quantity) {
+//        this(productId, productPrice, productName, productSnapshotDate, productType, quantity);
+//    }
 
     public OfferItem(String productId, BigDecimal productPrice, String productName, Date productSnapshotDate,
             String productType, int quantity) {
-        this(productId, productPrice, productName, productSnapshotDate, productType, quantity, null, null);
-    }
-
-    public OfferItem(String productId, BigDecimal productPrice, String productName, Date productSnapshotDate,
-            String productType, int quantity, BigDecimal discount, String discountCause) {
-        this.productId = productId;
-        this.productPrice = productPrice;
-        this.productName = productName;
-        this.productSnapshotDate = productSnapshotDate;
-        this.productType = productType;
+//        this.productId = productId;
+//        this.productPrice = productPrice;
+//        this.productName = productName;
+//        this.productSnapshotDate = productSnapshotDate;
+//        this.productType = productType;
 
         this.quantity = quantity;
-        this.discount = discount;
-        this.discountCause = discountCause;
+//        this.discount = discount;
+//        this.discountCause = discountCause;
 
         BigDecimal discountValue = new BigDecimal(0);
-        if (discount != null) {
-            discountValue = discountValue.subtract(discount);
-        }
+//        if (discount != null) {
+//            discountValue = discountValue.subtract(discount);
+//        }
 
         this.totalCost = productPrice.multiply(new BigDecimal(quantity)).subtract(discountValue);
     }
 
-    public String getProductId() {
-        return productId;
-    }
-
-    public BigDecimal getProductPrice() {
-        return productPrice;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public Date getProductSnapshotDate() {
-        return productSnapshotDate;
-    }
-
-    public String getProductType() {
-        return productType;
-    }
+//    public String getProductId() {
+//        return productId;
+//    }
+//
+//    public BigDecimal getProductPrice() {
+//        return productPrice;
+//    }
+//
+//    public String getProductName() {
+//        return productName;
+//    }
+//
+//    public Date getProductSnapshotDate() {
+//        return productSnapshotDate;
+//    }
+//
+//    public String getProductType() {
+//        return productType;
+//    }
 
     public BigDecimal getTotalCost() {
         return totalCost;
     }
 
-    public String getTotalCostCurrency() {
-        return currency;
-    }
+//    public String getTotalCostCurrency() {
+//        return currency;
+//    }
 
-    public BigDecimal getDiscount() {
-        return discount;
-    }
-
-    public String getDiscountCause() {
-        return discountCause;
-    }
+//    public BigDecimal getDiscount() {
+//        return discount;
+//    }
+//
+//    public String getDiscountCause() {
+//        return discountCause;
+//    }
 
     public int getQuantity() {
         return quantity;
@@ -108,11 +119,11 @@ public class OfferItem {
     public int hashCode() {
         final int prime = 31;
         int result = 1;
-        result = prime * result + (discount == null ? 0 : discount.hashCode());
-        result = prime * result + (productName == null ? 0 : productName.hashCode());
-        result = prime * result + (productPrice == null ? 0 : productPrice.hashCode());
-        result = prime * result + (productId == null ? 0 : productId.hashCode());
-        result = prime * result + (productType == null ? 0 : productType.hashCode());
+//        result = prime * result + (discount == null ? 0 : discount.hashCode());
+//        result = prime * result + (productName == null ? 0 : productName.hashCode());
+//        result = prime * result + (productPrice == null ? 0 : productPrice.hashCode());
+//        result = prime * result + (productId == null ? 0 : productId.hashCode());
+//        result = prime * result + (productType == null ? 0 : productType.hashCode());
         result = prime * result + quantity;
         result = prime * result + (totalCost == null ? 0 : totalCost.hashCode());
         return result;
@@ -130,37 +141,37 @@ public class OfferItem {
             return false;
         }
         OfferItem other = (OfferItem) obj;
-        if (discount == null) {
-            if (other.discount != null) {
-                return false;
-            }
-        } else if (!discount.equals(other.discount)) {
-            return false;
-        }
-        if (productName == null) {
-            if (other.productName != null) {
-                return false;
-            }
-        } else if (!productName.equals(other.productName)) {
-            return false;
-        }
-        if (productPrice == null) {
-            if (other.productPrice != null) {
-                return false;
-            }
-        } else if (!productPrice.equals(other.productPrice)) {
-            return false;
-        }
-        if (productId == null) {
-            if (other.productId != null) {
-                return false;
-            }
-        } else if (!productId.equals(other.productId)) {
-            return false;
-        }
-        if (productType != other.productType) {
-            return false;
-        }
+//        if (discount == null) {
+//            if (other.discount != null) {
+//                return false;
+//            }
+//        } else if (!discount.equals(other.discount)) {
+//            return false;
+//        }
+//        if (productName == null) {
+//            if (other.productName != null) {
+//                return false;
+//            }
+//        } else if (!productName.equals(other.productName)) {
+//            return false;
+//        }
+//        if (productPrice == null) {
+//            if (other.productPrice != null) {
+//                return false;
+//            }
+//        } else if (!productPrice.equals(other.productPrice)) {
+//            return false;
+//        }
+//        if (productId == null) {
+//            if (other.productId != null) {
+//                return false;
+//            }
+//        } else if (!productId.equals(other.productId)) {
+//            return false;
+//        }
+//        if (productType != other.productType) {
+//            return false;
+//        }
         if (quantity != other.quantity) {
             return false;
         }
@@ -176,36 +187,36 @@ public class OfferItem {
 
     /**
      *
-     * @param item
+//     * @param item
      * @param delta
      *            acceptable percentage difference
      * @return
      */
     public boolean sameAs(OfferItem other, double delta) {
-        if (productName == null) {
-            if (other.productName != null) {
-                return false;
-            }
-        } else if (!productName.equals(other.productName)) {
-            return false;
-        }
-        if (productPrice == null) {
-            if (other.productPrice != null) {
-                return false;
-            }
-        } else if (!productPrice.equals(other.productPrice)) {
-            return false;
-        }
-        if (productId == null) {
-            if (other.productId != null) {
-                return false;
-            }
-        } else if (!productId.equals(other.productId)) {
-            return false;
-        }
-        if (productType != other.productType) {
-            return false;
-        }
+//        if (productName == null) {
+//            if (other.productName != null) {
+//                return false;
+//            }
+//        } else if (!productName.equals(other.productName)) {
+//            return false;
+//        }
+//        if (productPrice == null) {
+//            if (other.productPrice != null) {
+//                return false;
+//            }
+//        } else if (!productPrice.equals(other.productPrice)) {
+//            return false;
+//        }
+//        if (productId == null) {
+//            if (other.productId != null) {
+//                return false;
+//            }
+//        } else if (!productId.equals(other.productId)) {
+//            return false;
+//        }
+//        if (productType != other.productType) {
+//            return false;
+//        }
 
         if (quantity != other.quantity) {
             return false;
