+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,0 +1,13 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.util.Date;
+
+public class Product {
+    private String productId;
+    private String productType;
+    private String productName;
+    private Date productSnapshotDate;
+    private Money price;
+
+
+}
